+++ Tests/Editor/SmokeTest.cs
@@ -1,0 +1,13 @@
+using NUnit.Framework;
+
+namespace skner.DualGrid.Tests
+{
+    public class SmokeTest
+    {
+        [Test]
+        public void TestFrameworkWorks()
+        {
+            Assert.AreEqual(2, 1 + 1);
+        }
+    }
+}